Move impl ToTokens to codegen/rust
