+++ espr/src/semantics/type_decl.rs
@@ -1,110 +1,36 @@
 use super::*;
 use crate::ast;
-use inflector::Inflector;
 
 /// Rename of primitive type,
 /// e.g. `TYPE label = STRING; ENDTYPE;`
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Simple {
-    id: String,
-    ty: SimpleType,
-}
-
-impl ToTokens for Simple {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        let id = format_ident!("{}", &self.id.to_pascal_case());
-        let ty = &self.ty;
-        tokens.append_all(quote! {
-            pub type #id = #ty;
-        });
-    }
+    pub id: String,
+    pub ty: SimpleType,
 }
 
 /// Rename of user defined type,
 /// e.g. `TYPE box_height = positive_ratio_measure; END_TYPE;`
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Rename {
-    id: String,
-    ty: TypeRef,
-}
-
-impl ToTokens for Rename {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        let id = format_ident!("{}", &self.id.to_pascal_case());
-        let ty = &self.ty;
-        tokens.append_all(quote! {
-            pub type #id = #ty;
-        });
-    }
+    pub id: String,
+    pub ty: TypeRef,
 }
 
 /// Enumeration of values,
 /// e.g. `TYPE text_path = ENUMERATION OF (up, right, down, left); END_TYPE;`
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Enumeration {
-    id: String,
-    items: Vec<String>,
-}
-
-impl ToTokens for Enumeration {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        let id = format_ident!("{}", &self.id.to_pascal_case());
-        let items: Vec<_> = self
-            .items
-            .iter()
-            .map(|i| format_ident!("{}", i.to_pascal_case()))
-            .collect();
-        tokens.append_all(quote! {
-            #[derive(Debug, Clone)]
-            pub enum #id {
-                #( #items ),*
-            }
-        });
-    }
+    pub id: String,
+    pub items: Vec<String>,
 }
 
 /// Select of user defined types,
 /// e.g. `TYPE geometric_set_select = SELECT (point, curve); END_TYPE;`
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Select {
-    id: String,
-    types: Vec<TypeRef>,
-}
-
-impl ToTokens for Select {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        let id = format_ident!("{}", &self.id.to_pascal_case());
-        let mut entries = Vec::new();
-        let mut entry_types = Vec::new();
-        for ty in &self.types {
-            match ty {
-                TypeRef::Entity {
-                    name,
-                    has_supertype_decl,
-                    ..
-                } => {
-                    let name = format_ident!("{}", name.to_pascal_case());
-                    entries.push(quote! { #name });
-                    if *has_supertype_decl {
-                        // avoid Box<Box<XxxAny>>
-                        entry_types.push(quote! { #ty });
-                    } else {
-                        entry_types.push(quote! { Box<#ty> });
-                    }
-                }
-                _ => {
-                    entries.push(ty.to_token_stream());
-                    entry_types.push(quote! { Box<#ty> });
-                }
-            }
-        }
-        tokens.append_all(quote! {
-            #[derive(Debug, Clone)]
-            pub enum #id {
-                #(#entries(#entry_types)),*
-            }
-        });
-    }
+    pub id: String,
+    pub types: Vec<TypeRef>,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -152,16 +78,5 @@ impl Legalize for TypeDecl {
             }
             _ => panic!(),
         })
-    }
-}
-
-impl ToTokens for TypeDecl {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        match self {
-            TypeDecl::Simple(simple) => simple.to_tokens(tokens),
-            TypeDecl::Rename(rename) => rename.to_tokens(tokens),
-            TypeDecl::Enumeration(e) => e.to_tokens(tokens),
-            TypeDecl::Select(select) => select.to_tokens(tokens),
-        }
     }
 }
